Use separate pages for doc list views
