+++ layouts/page.vue
@@ -1,6 +1,6 @@
 <template>
   <CoreScrollable>
-    <router-view :inheritAttrs="true" class="flex"></router-view>
+    <router-view :inheritAttrs="true" class="container mx-auto p-4 md:px-16 md:py-8"></router-view>
   </CoreScrollable>
 </template>
